+++ py_basic_commands/file_dir_scripts/create_dirs.py
@@ -21,7 +21,7 @@ class CreateDirs(Base):
             - `bool`: If all directories were created or not"""
 
             # Check input values
-            do_print = self._check_input_val(do_print, self._do_print)
+            do_print = self._check_input_val(do_print, self.do_print)
 
             fprint.config(do_print=do_print)