+++ py_basic_commands/file_dir_scripts/create_file.py
@@ -51,7 +51,7 @@ class CreateFile(Base):
 
         # Check input values
         force = self._check_input_val(force, self._force)
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)