+++ py_basic_commands/file_dir_scripts/get_src_path.py
@@ -32,7 +32,7 @@ class GetSourcePath(Base):
 
         # Check input values
         ret_val = self._check_input_val(ret_val, self._ret_val)
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)