+++ py_basic_commands/file_dir_scripts/join_path.py
@@ -36,7 +36,7 @@ class JoinPath(Base):
         join_with = self._check_input_val(join_with, self._join_with)
         remove_empty = self._check_input_val(remove_empty, self._remove_empty)
         dir_end = self._check_input_val(dir_end, self._dir_end)
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)