+++ py_basic_commands/file_dir_scripts/read_file.py
@@ -74,7 +74,7 @@ class ReadFile(Base):
         do_strip    = self._check_input_val(do_strip, self._do_strip)
         do_lower    = self._check_input_val(do_lower, self._do_lower)
         encoding    = self._check_input_val(encoding, self._encoding)
-        do_print    = self._check_input_val(do_print, self._do_print)
+        do_print    = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)