+++ py_basic_commands/file_dir_scripts/remove_file_dir.py
@@ -31,7 +31,7 @@ class RemoveFileDir(Base):
 
         # Check input values
         force    = self._check_input_val(force, self._force)
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         if do == 'd': # Directory
             try: