+++ py_basic_commands/json_scripts/create_json.py
@@ -29,7 +29,7 @@ class CreateJson(Base):
 
         # Check input values
         force = self._check_input_val(force, self._force)
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)