+++ py_basic_commands/json_scripts/read_json.py
@@ -25,7 +25,7 @@ class ReadJson(Base):
         """
 
         # Check input values
-        do_print = self._check_input_val(do_print, self._do_print)
+        do_print = self._check_input_val(do_print, self.do_print)
 
         fprint.config(do_print=do_print)